requirements.txt
pygame>=2.1
Pillow>=9.0
pymunk>=6.0
